requirements.txt
# requirements.txt
# torch>=2.2.0
# torchvision>=0.17.0
transformers>=4.49.0
accelerate>=0.28.0
peft>=0.10.0
datasets>=2.18.0
trl>=0.8.1
bitsandbytes>=0.43.0
sentencepiece>=0.2.0
protobuf>=4.25.3
boto3>=1.34.0
sagemaker>=2.214.0
pillow>=10.2.0
httpx>=0.27.0
timm
einops